refactor: moved shard utils to general utils location
